product_tv.mk: set persist.sys.hdmi.keep_awake=false
This keeps the device from suspending and turning of HDMI if the connected display is turned off
with HDMI CEC enabled. This is to avoid an issue where HDMI is not enabled again when the display
is powered on again (seen with some displays). When we have proper HDMI CEC configuration working
this should possibly be reverted.